refactor: refactor out read and write functions into new files
